requirements.txt
Flask==2.2.2
psycopg2==2.9.3
python-dotenv==0.21.0
pytest
pytest-cov
pymongo==3.1
requests==2.26.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-PyMongo==3.0.0
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2